refactor: separate postprocessing and runtime inference logic
- Move postprocessing functions from runtime.py to utils/vision.py
- Update runtime classes to return raw model outputs
- Modify local and remote model classes to apply postprocessing separately
- Improve logging with more detailed inference time and detection count information
- Enhance error handling for cases with no detections
