+++ tests/utils/test_vision.py
@@ -4,18 +4,22 @@ import math
 import numpy as np
 import supervision as sv
 
-from focoos.ports import FocoosDet
+from focoos.ports import FocoosDet, FocoosTask
 from focoos.utils.vision import (
     base64mask_to_mask,
     binary_mask_to_base64,
     class_to_index,
+    det_postprocess,
     fai_detections_to_sv,
+    get_postprocess_fn,
     image_loader,
     image_preprocess,
     index_to_class,
-    mask_to_xyxy,
+    instance_postprocess,
+    masks_to_xyxy,
     scale_detections,
     scale_mask,
+    semseg_postprocess,
     sv_to_fai_detections,
 )
 
@@ -180,28 +184,144 @@ def test_sv_to_focoos_detections(sv_detections: sv.Detections):
     assert isinstance(result_focoos_detection.mask, str), "Mask should be a string"
 
 
-def test_mask_to_xyxy():
+def test_masks_to_xyxy():
     # Basic case: a single mask with one active pixel
     mask1 = np.zeros((1, 5, 5), dtype=bool)
     mask1[0, 2, 3] = True  # One active pixel at (2,3)
-    assert np.array_equal(mask_to_xyxy(mask1), np.array([[3, 2, 3, 2]]))
+    assert np.array_equal(masks_to_xyxy(mask1), np.array([[3, 2, 3, 2]]))
 
     # Case with a rectangle
     mask2 = np.zeros((1, 5, 5), dtype=bool)
     mask2[0, 1:4, 2:5] = True  # Rectangle between (1,2) and (3,4)
-    assert np.array_equal(mask_to_xyxy(mask2), np.array([[2, 1, 4, 3]]))
+    assert np.array_equal(masks_to_xyxy(mask2), np.array([[2, 1, 4, 3]]))
 
     # Case with multiple masks
     masks = np.zeros((2, 5, 5), dtype=bool)
     masks[0, 1:4, 2:5] = True  # First rectangle
     masks[1, 0:3, 1:4] = True  # Second rectangle
     expected = np.array([[2, 1, 4, 3], [1, 0, 3, 2]])
-    assert np.array_equal(mask_to_xyxy(masks), expected)
-
-    # Case with an empty mask
-    empty_mask = np.zeros((1, 5, 5), dtype=bool)
-    assert np.array_equal(mask_to_xyxy(empty_mask), np.array([[0, 0, 0, 0]]))
+    assert np.array_equal(masks_to_xyxy(masks), expected)
 
     # Case with a mask covering the entire image
     full_mask = np.ones((1, 5, 5), dtype=bool)
-    assert np.array_equal(mask_to_xyxy(full_mask), np.array([[0, 0, 4, 4]]))
+    assert np.array_equal(masks_to_xyxy(full_mask), np.array([[0, 0, 4, 4]]))
+
+
+def test_det_post_process():
+    cls_ids = np.array([1, 2, 3])
+    boxes = np.array([[0.1, 0.2, 0.3, 0.4], [0.5, 0.6, 0.7, 0.8], [0.9, 1.0, 1.1, 1.2]])
+    confs = np.array([0.8, 0.9, 0.7])
+    out = [cls_ids, boxes, confs]
+
+    im0_shape = (640, 480)
+    conf_threshold = 0.75
+    sv_detections = det_postprocess(out, im0_shape, conf_threshold)
+
+    np.testing.assert_array_equal(sv_detections.xyxy, np.array([[48, 128, 144, 256], [240, 384, 336, 512]]))
+    assert sv_detections.class_id is not None
+    np.testing.assert_array_equal(sv_detections.class_id, np.array([1, 2]))
+    assert sv_detections.confidence is not None
+    np.testing.assert_array_equal(sv_detections.confidence, np.array([0.8, 0.9]))
+
+
+def test_semseg_postprocess():
+    cls_ids = np.array([1, 2, 3])
+    mask = np.array(
+        [
+            [0, 1, 1, 2],
+            [0, 1, 2, 2],
+            [0, 0, 1, 2],
+        ]
+    )
+    confs = np.array([0.7, 0.9, 0.8])
+    out = [
+        np.expand_dims(cls_ids, axis=0),
+        np.expand_dims(mask, axis=0),
+        np.expand_dims(confs, axis=0),
+    ]
+
+    im0_shape = (3, 4)
+    conf_threshold = 0.75
+
+    sv_detections = semseg_postprocess(out, im0_shape, conf_threshold)
+
+    # Expected masks
+    expected_masks = np.array(
+        [
+            [
+                [False, True, True, False],
+                [False, True, False, False],
+                [False, False, True, False],
+            ],  # Class 1
+            [
+                [False, False, False, True],
+                [False, False, True, True],
+                [False, False, False, True],
+            ],  # Class 2
+        ]
+    )
+
+    # Assertions
+    assert sv_detections.mask is not None
+    np.testing.assert_array_equal(sv_detections.mask, expected_masks)
+    assert sv_detections.class_id is not None
+    np.testing.assert_array_equal(sv_detections.class_id, np.array([2, 3]))
+    assert sv_detections.confidence is not None
+    np.testing.assert_array_equal(sv_detections.confidence, np.array([0.9, 0.8]))
+    assert sv_detections.xyxy.shape == (2, 4)
+
+
+def test_get_postprocess_fn():
+    """
+    Test the get_postprocess_fn function to ensure it returns
+    the correct postprocessing function for each task.
+    """
+    # Test detection task
+    det_fn = get_postprocess_fn(FocoosTask.DETECTION)
+    assert det_fn == det_postprocess, "Detection task should return det_postprocess function"
+
+    # Test instance segmentation task
+    instance_fn = get_postprocess_fn(FocoosTask.INSTANCE_SEGMENTATION)
+    assert instance_fn == instance_postprocess, "Instance segmentation task should return instance_postprocess function"
+
+    # Test semantic segmentation task
+    semseg_fn = get_postprocess_fn(FocoosTask.SEMSEG)
+    assert semseg_fn == semseg_postprocess, "Semantic segmentation task should return semseg_postprocess function"
+
+    # Test all FocoosTask values to ensure no exceptions
+    for task in FocoosTask:
+        fn = get_postprocess_fn(task)
+        assert callable(fn), f"Postprocess function for {task} should be callable"
+
+
+def test_instance_postprocess():
+    """Test instance segmentation postprocessing"""
+    cls_ids = np.array([0, 1])
+    masks = np.zeros((2, 100, 100))
+    masks[0, 10:30, 10:30] = 1
+    masks[1, 40:60, 40:60] = 1
+    confs = np.array([0.95, 0.85])
+    out = [[cls_ids], [masks], [confs]]
+
+    result = instance_postprocess(out, (100, 100), 0.8)
+
+    assert isinstance(result, sv.Detections)
+    assert len(result) == 2
+    assert result.mask is not None
+    assert result.xyxy is not None
+    assert result.class_id is not None
+    assert result.confidence is not None
+
+
+def test_confidence_threshold_filtering():
+    """Test that confidence threshold filtering works correctly"""
+    out = [
+        np.array([0, 1, 2]),  # cls_ids
+        np.array([[0.1, 0.1, 0.3, 0.3], [0.4, 0.4, 0.6, 0.6], [0.7, 0.7, 0.9, 0.9]]),  # boxes
+        np.array([0.95, 0.55, 0.85]),  # confs
+    ]
+
+    result = det_postprocess(out, (100, 100), conf_threshold=0.8)
+
+    assert len(result) == 2  # Should only keep detections with conf > 0.8
+    assert all(conf > 0.8 for conf in result.confidence)
